renaming the core module to common. Adjusting the final output jar names and versions
