+++ common/build.gradle
@@ -4,8 +4,8 @@ plugins {
     id 'org.parchmentmc.librarian.forgegradle' version '1.+'
 }
 
-group = 'com.dudenduke.hudalternatives'
-version = '1.0'
+group = mod_group_id
+version = module_version + "-mc" + minecraft_version
 
 java {
     toolchain.languageVersion = JavaLanguageVersion.of(17)
@@ -44,4 +44,9 @@ dependencies {
 
 test {
     useJUnitPlatform()
+}
+
+jar {
+    archiveBaseName=mod_group_name
+    archiveAppendix=project.name
 }
