+++ common/gradle.properties
@@ -1,0 +1,1 @@
+module_version=1.1.0
