+++ common/src/main/java/com/dudenduke/hudalternatives/common/Constants.java
@@ -1,4 +1,4 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 public class Constants {
 	private static final String modParent = "hudalternatives";
