+++ common/src/main/java/com/dudenduke/hudalternatives/common/Dimensions.java
@@ -1,3 +1,3 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 public record Dimensions(int width, int height) { }
