+++ common/src/main/java/com/dudenduke/hudalternatives/common/LivingVehicleType.java
@@ -1,4 +1,4 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 public enum LivingVehicleType {
     Horse, Donkey, Mule, Pig, Camel, Skeleton_horse, Strider, Unknown
