+++ common/src/main/java/com/dudenduke/hudalternatives/common/SurvivalPlayerSnapshot.java
@@ -1,4 +1,4 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.tags.BlockTags;
