+++ common/src/main/java/com/dudenduke/hudalternatives/common/Vector2.java
@@ -1,0 +1,3 @@
+package com.dudenduke.hudalternatives.common;
+
+public record Vector2 (int x, int y) { }
