+++ common/src/main/java/com/dudenduke/hudalternatives/common/Vector2F.java
@@ -1,3 +1,3 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 public record Vector2F (float x, float y) { }
