+++ common/src/main/java/com/dudenduke/hudalternatives/common/WeaponType.java
@@ -1,4 +1,4 @@
-package com.dudenduke.hudalternatives.core;
+package com.dudenduke.hudalternatives.common;
 
 public enum WeaponType {
     ANY, SWORD, AXE, MELEE, RANGED
