+++ common/src/main/resources/assets/minecraft/textures/gui/icons.png
