+++ common/src/main/resources/data/forge/tags/items/ranged_weapons.json
