+++ common/src/main/resources/data/forge/tags/items/weapons.json
