--- core/src/main/java/com/dudenduke/hudalternatives/core/Vector2.java
@@ -1,3 +1,0 @@
-package com.dudenduke.hudalternatives.core;
-
-public record Vector2 (int x, int y) { }
