+++ gradle.properties
@@ -3,6 +3,12 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
+# Hud Alternatives
+mod_group_name=hud-alternatives
+# The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
+# This should match the base package used for the mod sources.
+# See https://maven.apache.org/guides/mini/guide-naming-conventions.html
+mod_group_id=com.dudenduke.hudalternatives
 
 ## Environment Properties
