+++ minimalmodern/gradle.properties
@@ -9,10 +9,6 @@ mod_name=Minimal Modern
 mod_license=All Rights Reserved
 # The mod version. See https://semver.org/
 mod_version=1.2.0
-# The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
-# This should match the base package used for the mod sources.
-# See https://maven.apache.org/guides/mini/guide-naming-conventions.html
-mod_group_id=com.dudenduke.hudalternatives
 # The authors of the mod. This is a simple text string that is used for display purposes in the mod list.
 mod_authors=Charles Haines, Ryan Conklin
 # The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
